requirements.txt
flask==1.1.2
pyaudio
google-cloud-pubsub
python-socketio
